Use the IMessageDialogService in mock to verify the delete dialog is called and add tests

+++ FriendStorage.UI/ViewModel/FriendEditViewModel.cs
@@ -73,8 +73,12 @@ namespace FriendStorage.UI.ViewModel
 
         private void OnDeleteExecute(object obj)
         {
-            _dataProvider.DeleteFriend(Friend.Id);
-            _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            var result = _messageDialogService.ShowYesNoDialog("Delete friend", "");
+            if (result == MessageDialogResult.Yes)
+            {
+                _dataProvider.DeleteFriend(Friend.Id);
+                _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            }
         }
 
         private bool OnSaveCanExecute(object arg)
